fiber_photometry/requirements.txt
numpy
scipy
pandas
matplotlib
tdt
PyYAML
